+++ .jenkins/short-perf-test-gpu.sh
@@ -15,7 +15,11 @@ echo "Running GPU perf test for PyTorch..."
 export PYTORCH_COMMIT_ID=$(git log --format="%H" -n 1)
 
 # Get baseline file from https://github.com/yf225/perf-tests
-cp /var/lib/jenkins/host-workspace/perf_test_numbers_gpu.json perf_test_numbers_gpu.json
+if [ -f /var/lib/jenkins/host-workspace/perf_test_numbers_gpu.json ]; then
+    cp /var/lib/jenkins/host-workspace/perf_test_numbers_gpu.json perf_test_numbers_gpu.json
+else
+    curl https://raw.githubusercontent.com/yf225/perf-tests/master/perf_test_numbers_gpu.json -O
+fi
 
 if [[ "$GIT_COMMIT" == *origin/master* ]]; then
     # Prepare new baseline file
